Перенесены модели Dto
